+++ app/kumactl/pkg/ca/client.go
@@ -2,6 +2,7 @@ package ca
 
 import (
 	"bytes"
+	tls1 "crypto/tls"
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
@@ -38,7 +39,8 @@ func NewProvidedCaClient(address string, config *kumactl_config.Context_AdminApi
 		return nil, errors.Wrapf(err, "failed to parse the server URL")
 	}
 	httpClient := &http.Client{
-		Timeout: timeout,
+		Timeout:   timeout,
+		Transport: &http.Transport{TLSClientConfig: &tls1.Config{InsecureSkipVerify: true}},
 	}
 	if baseURL.Scheme == "https" {
 		if !config.HasClientCert() {