+++ pkg/api-server/config_ws_test.go
@@ -49,6 +49,17 @@ var _ = Describe("Config WS", func() {
 		json := fmt.Sprintf(`
         {
           "apiServer": {
+            "catalog": {
+              "bootstrap": {
+                "url": ""
+              },
+              "monitoringAssignment": {
+                "url": ""
+              },
+              "sds": {
+                "url": ""
+              }
+            },
             "corsAllowedDomains": [
               ".*"
             ],
@@ -159,7 +170,9 @@ var _ = Describe("Config WS", func() {
             "dataplaneConfigurationRefreshInterval": "1s",
             "dataplaneStatusFlushInterval": "1s",
             "diagnosticsPort": 5680,
-            "grpcPort": 5678
+            "grpcPort": 5678,
+            "tlsCertFile": "",
+            "tlsKeyFile": ""
           }
         }
 		`, port)