+++ pkg/xds/bootstrap/template.go
@@ -13,6 +13,7 @@ type configParameters struct {
 	XdsConnectTimeout  time.Duration
 	AccessLogPipe      string
 	DataplaneTokenPath string
+	CertBytes          string
 }
 
 const configTemplate string = `
@@ -77,6 +78,19 @@ static_resources:
               socket_address:
                 address: {{ .XdsHost }}
                 port_value: {{ .XdsPort }}
+{{ if .CertBytes }}
+    transport_socket:
+      name: envoy.transport_sockets.tls
+      typed_config:
+        "@type": type.googleapis.com/envoy.api.v2.auth.UpstreamTlsContext
+        sni: {{ .XdsHost }}
+        common_tls_context:
+          tls_params:
+            tls_minimum_protocol_version: TLSv1_2
+          validation_context:
+            trusted_ca:
+              inline_bytes: "{{ .CertBytes }}"
+{{ end }}
   - name: access_log_sink
     connect_timeout: {{ .XdsConnectTimeout }}
     type: STATIC