+++ pkg/xds/server/grpc.go
@@ -6,7 +6,9 @@ import (
 
 	envoy_discovery "github.com/envoyproxy/go-control-plane/envoy/service/discovery/v2"
 	envoy_xds "github.com/envoyproxy/go-control-plane/pkg/server"
+	"github.com/pkg/errors"
 	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials"
 
 	"github.com/Kong/kuma/pkg/core"
 	"github.com/Kong/kuma/pkg/core/runtime/component"
@@ -19,8 +21,10 @@ var (
 )
 
 type grpcServer struct {
-	server envoy_xds.Server
-	port   int
+	server      envoy_xds.Server
+	port        int
+	tlsCertFile string
+	tlsKeyFile  string
 }
 
 // Make sure that grpcServer implements all relevant interfaces
@@ -31,6 +35,14 @@ var (
 func (s *grpcServer) Start(stop <-chan struct{}) error {
 	var grpcOptions []grpc.ServerOption
 	grpcOptions = append(grpcOptions, grpc.MaxConcurrentStreams(grpcMaxConcurrentStreams))
+	useTLS := s.tlsCertFile != ""
+	if useTLS {
+		creds, err := credentials.NewServerTLSFromFile(s.tlsCertFile, s.tlsKeyFile)
+		if err != nil {
+			return errors.Wrap(err, "failed to load TLS certificate")
+		}
+		grpcOptions = append(grpcOptions, grpc.Creds(creds))
+	}
 	grpcServer := grpc.NewServer(grpcOptions...)
 
 	lis, err := net.Listen("tcp", fmt.Sprintf(":%d", s.port))
@@ -51,7 +63,7 @@ func (s *grpcServer) Start(stop <-chan struct{}) error {
 			grpcServerLog.Info("terminated normally")
 		}
 	}()
-	grpcServerLog.Info("starting", "interface", "0.0.0.0", "port", s.port)
+	grpcServerLog.Info("starting", "interface", "0.0.0.0", "port", s.port, "tls", useTLS)
 
 	select {
 	case <-stop: