+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
 ## master
+* feat kuma-dp and kumactl can communiate with kuma-cp over https
+  [#633](https://github.com/Kong/kuma/pull/633)
 * feat(kuma-cp) envoy configs for fault injections
   [#649](https://github.com/Kong/kuma/pull/649)
 * feat: endpoints for fetching resources from all meshes 
@@ -43,7 +45,8 @@
   [#622](https://github.com/Kong/kuma/pull/622)
 * feature: validate `<port>.service.kuma.io/protocol` annotations on K8S Service objects
   [#611](https://github.com/Kong/kuma/pull/611)
-
+* feat: added `kumactl get` command for individual resources
+  [#667](https://github.com/Kong/kuma/pull/667)
 ## [0.4.0]
 
 > Released on 2020/02/28